% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{R_LM_Isotope_Companion}
\alias{R_LM_Isotope_Companion}
\title{Launch the R-LM-Isotope-Companion addin}
\usage{
R_LM_Isotope_Companion()
}
\value{
This function launches an interactive RStudio gadget.
}
\description{
Opens an RStudio gadget for querying a local language model served by
LM Studio about R code.
}
\examples{
\dontrun{
R_LM_Isotope_Companion()
}
}
